// src/csvdataintegrationapp.client/src/app/services/models/employee.ts
export class EmployeeModel {
    address?: string | null;
    address2?: string | null;
    dateOfBirth: Date = new Date();
    emailHome?: string | null;
    forenames?: string | null;
    id?: string;
    mobile?: string | null;
    payrollNumber?: string | null;
    postcode?: string | null;
    startDate: Date = new Date();
    surname?: string | null;
    telephone?: string | null;
  }